refactor(backend): backlog — service extraction, validation, error handling, N+1
A1 — extract data/aggregation logic out of controllers into services:
- servers_controller.paginate -> new ServerListingService.paginate
- advertisements_controller.stats -> new AdvertisementStatsService.getStats
  (logic moved verbatim; same queries, output shapes, caching).

A2 — validate external input with VineJS:
- server_categories_controller store/destroy -> attachServerCategoryValidator
- categories_controller destroy/update -> validated id
- posts_controller.previewPlaceholder -> validated placeholderName/serverId
  (invalid input now returns 422 instead of proceeding).

A6 — stop swallowing/normalizing errors:
- stats_controller index/globalStats: drop the catch-all so AdonisJS formats
  validation (422) and HTTP errors; generic errors yield a proper 500.
- servers_controller: remove try/catch around pure CSV parsing that silently
  dropped category/language filters.
- placeholder_service: no more bare catch {} swallowing DB errors.

A4 — fix the N+1 in placeholder_service.preloadServerData: ~6*N per-server
queries replaced with set-based queries (whereIn + DISTINCT ON per-server
latest/first/peak, one grouped query for avg/median). Missing servers map to
exists:false; DB errors propagate.

backend typecheck, lint and build are green. NEEDS RUNTIME QA (DB): the N+1
rewrite's aggregates and the validation/error-shape changes are behavior-
affecting and were not runtime-tested.
