+++ backend/app/controllers/servers_controller.ts
@@ -11,7 +11,7 @@ import {
 import type { NormalizedPing } from '../../minecraft-ping/minecraft_ping.js'
 import Server from '../models/server.js'
 import CacheService from '#services/cache_service'
-import StatsService from '#services/stat_service'
+import ServerListingService from '#services/server_listing_service'
 import DuplicateDetectionService from '#services/duplicate_detection_service'
 import Language from '#models/language'
 
@@ -315,7 +315,7 @@ export default class ServersController {
       cacheKey,
       ttl,
       async () => {
-        const result = await this.runPaginateQuery({
+        const result = await ServerListingService.paginate({
           page,
           limit,
           categoryIds,
@@ -356,137 +356,5 @@ export default class ServersController {
       if (ids.length >= ServersController.MAX_IDS) break
     }
     return ids
-  }
-
-  private async runPaginateQuery(opts: {
-    page: number
-    limit: number
-    categoryIds?: string
-    languageIds?: string
-    search: string
-    type?: 'java' | 'bedrock'
-    ids: number[]
-  }) {
-    const { page, limit, categoryIds, languageIds, search, type, ids } = opts
-
-    // Ordering : on ne classe par `last_player_count` que si le dernier ping
-    // réussi est récent (< 30 min, soit ~3 cycles de 10 min). Sinon le serveur
-    // est traité comme "stale" et descend en bas, peu importe son ancien count.
-    // Évite qu'un serveur down depuis des jours reste top juste parce qu'il
-    // avait 500 joueurs avant de tomber.
-    let query = Server.query()
-      .preload('user', (userQuery) => userQuery.select('id', 'username', 'avatarUrl'))
-      .preload('categories')
-      .preload('growthStat')
-      .preload('languages')
-
-    // Restriction à une liste explicite d'IDs (section favoris) : `whereIn` avec
-    // bindings, donc safe même si les IDs venaient d'une source moins fiable.
-    if (ids.length > 0) {
-      query = query.whereIn('id', ids)
-    }
-
-    query = query
-      .orderByRaw(
-        `CASE
-          WHEN last_stats_at > now() - interval '30 minutes'
-            THEN COALESCE(last_player_count, -1)
-          ELSE -1
-         END DESC`
-      )
-      .orderBy('last_stats_at', 'desc')
-
-    if (search) {
-      query = query.where((builder) => {
-        builder.whereILike('name', `%${search}%`).orWhereILike('address', `%${search}%`)
-      })
-    }
-
-    if (type) {
-      query = query.where('type', type)
-    }
-
-    if (categoryIds) {
-      try {
-        const categoryIdList = categoryIds
-          .split(',')
-          .map((id: string) => Number.parseInt(id.trim(), 10))
-          .filter((id: number) => !Number.isNaN(id))
-        if (categoryIdList.length > 0) {
-          query = query.whereHas('categories', (builder) => {
-            builder.whereIn('categories.id', categoryIdList)
-          })
-        }
-      } catch (error) {
-        console.error('Error processing categoryIds:', error)
-      }
-    }
-
-    if (languageIds) {
-      try {
-        const languageIdList = languageIds
-          .split(',')
-          .map((id: string) => Number.parseInt(id.trim(), 10))
-          .filter((id: number) => !Number.isNaN(id))
-        if (languageIdList.length > 0) {
-          query = query.whereHas('languages', (builder) => {
-            builder.whereIn('languages.id', languageIdList)
-          })
-        }
-      } catch (error) {
-        console.error('Error processing languageIds:', error)
-      }
-    }
-
-    const servers = await query.paginate(page, limit)
-
-    const now = Date.now()
-    const fromDate = now - 24 * 60 * 60 * 1000
-
-    // Important : Lucid SimplePaginator étend Array, donc `servers.map(...)` retourne
-    // une *nouvelle SimplePaginator* (via Symbol.species), pas un Array standard.
-    // Ce paginator a un `.toJSON()` qui réécrit la réponse en `{meta, data}` →
-    // double nesting côté HTTP. On extrait `servers.all()` (le rows[] réel) avant map.
-    const serverRows = servers.all()
-    const serverIds = serverRows.map((s) => s.id)
-    const statsByServer = await StatsService.getStatsBatch({
-      serverIds,
-      fromDate,
-      toDate: now,
-      interval: '1 hour',
-    })
-
-    const serversWithStats = serverRows.map((server) => {
-      const bucketed = (statsByServer.get(server.id) ?? []).map((row) => ({
-        serverId: server.id,
-        createdAt: row.created_at,
-        playerCount: row.player_count,
-        maxCount: row.max_count,
-      }))
-
-      const lastStat =
-        server.lastStatsAt !== null && server.lastPlayerCount !== null
-          ? [
-              {
-                serverId: server.id,
-                createdAt: server.lastStatsAt,
-                playerCount: server.lastPlayerCount,
-                maxCount: server.lastMaxCount,
-              },
-            ]
-          : []
-
-      return {
-        server,
-        stats: [...lastStat, ...bucketed],
-        categories: server.categories,
-        growthStat: server.growthStat,
-      }
-    })
-
-    return {
-      data: serversWithStats,
-      meta: servers.getMeta(),
-    }
   }
 }
