+++ backend/app/validators/server_category.ts
@@ -1,0 +1,8 @@
+import vine from '@vinejs/vine'
+
+export const attachServerCategoryValidator = vine.compile(
+  vine.object({
+    serverId: vine.number().positive(),
+    categoryId: vine.number().positive(),
+  })
+)
